move np_score to /np_score folder
