+++ src/core/libraries/np/np_score/np_score.cpp
@@ -4,7 +4,7 @@
 #include "common/logging/log.h"
 #include "core/libraries/error_codes.h"
 #include "core/libraries/libs.h"
-#include "core/libraries/np/np_score.h"
+#include "np_score.h"
 
 namespace Libraries::Np::NpScore {
