+++ src/core/libraries/np/np_score/np_score.h
